requirements.txt
fastapi
uvicorn
rhino3dm
python-dotenv
rhino3dm-compute
